requirements.txt
-e git+https://github.com/yaseen002/autodev-cli.git@46a0c0b80786abb8f7484a4a5fd0480a96486a7f#egg=autodev
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
gitdb==4.0.12
GitPython==3.1.45
hatchling==1.27.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
smmap==5.0.2
trove-classifiers==2025.9.11.17
typer==0.19.2
typing_extensions==4.15.0
urllib3==2.5.0
